// src/types/properyHomes.ts
export type PropertyHomes = {
  id: string;
  name_ar: string;
  name_en: string;
  operation: string;
  type: string;
  bedrooms: number;
  bathrooms: number;
  area: number;
  price: number;
  address_ar: string;
  address_en: string;
  details_ar: string;
  details_en: string;
  map_url: string;
  images: string[];
  created_at: string;
};